Create DigitalPort_Write.ino
This code reads a integer 2..13 from the serial port.
Then Toggles the output. All ports start as LOW
So, if you write 13, the port 13 will be set to HIGH the first time, and LOW the next.
A serial message will write to the console, what pin it high or low.